Add support for localization with English language
The Neerslag Card now supports localization. The first added language is English. The language as set by Home Assistant will be used automatically. If that is not available the browser language will be used. If the selected language is unavailable the Neerslag Card will default back to the Dutch language.